+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Changelog
-
--   Added a login function with Metamask
+-   Integrated login procedure with Ethererum Rinkeby testnet by using "register contract"
+-   Created a "register contract", which contains signup, getRegisteredAccount, deactivateAccount, transferOwnership function
+-   Added a "login function" with Metamask
 -   Added a basic layout consisting of a dashboard, a sidebar and a navbar